Put internal function raw type in own file
