+++ packages/codec/lib/format/errors.ts
@@ -18,6 +18,10 @@ import type * as Types from "./types";
 import type * as Ast from "@truffle/codec/ast/types";
 import type * as Storage from "@truffle/codec/storage/types";
 import type { PaddingType } from "@truffle/codec/common";
+import type {
+  FunctionInternalRawInfo,
+  FunctionInternalRawInfoPcPair
+} from "./internal";
 
 /*
  * SECTION 1: Generic types for values in general (including errors).
@@ -712,8 +716,7 @@ export interface FunctionInternalErrorResult {
  */
 export type FunctionInternalError =
   | FunctionInternalPaddingError
-  | NoSuchPcValueError
-  | NoSuchFunctionIndexError
+  | NoSuchInternalFunctionError
   | DeployedFunctionInConstructorError
   | MalformedInternalFunctionError;
 
@@ -732,29 +735,15 @@ export interface FunctionInternalPaddingError {
 }
 
 /**
- * Indicates that the function pointer being decoded (one given by PC value)
- * fails to point to a valid function, and also is not one of the default
- * values
+ * Indicates that the function pointer being decoded fails to point to a valid
+ * function, and also is not one of the default values
  *
  * @Category Function types
  */
-export interface NoSuchPcValueError {
-  kind: "NoSuchPcValueError";
+export interface NoSuchInternalFunctionError {
+  kind: "NoSuchInternalFunctionError";
   context: Types.ContractType;
-  deployedProgramCounter: number;
-  constructorProgramCounter: number;
-}
-
-/**
- * Indicates that the function pointer being decoded (one given by function
- * index) fails to point to a valid function, and also is not zero
- *
- * @Category Function types
- */
-export interface NoSuchFunctionIndexError {
-  kind: "NoSuchFunctionIndexError";
-  context: Types.ContractType;
-  functionIndex: number;
+  rawInformation: FunctionInternalRawInfo;
 }
 
 /**
@@ -766,8 +755,7 @@ export interface NoSuchFunctionIndexError {
 export interface DeployedFunctionInConstructorError {
   kind: "DeployedFunctionInConstructorError";
   context: Types.ContractType;
-  deployedProgramCounter: number;
-  constructorProgramCounter: number;
+  rawInformation: FunctionInternalRawInfoPcPair; //will never be index
 }
 
 /**
@@ -779,8 +767,7 @@ export interface DeployedFunctionInConstructorError {
 export interface MalformedInternalFunctionError {
   kind: "MalformedInternalFunctionError";
   context: Types.ContractType;
-  deployedProgramCounter: number;
-  constructorProgramCounter: number;
+  rawInformation: FunctionInternalRawInfoPcPair; //will never be index
 }
 
 /*
