+++ packages/codec/lib/format/values.ts
@@ -38,13 +38,15 @@ import type {
   ContractValueInfoUnknown,
   StringValueInfo
 } from "./elementary";
+import type { FunctionInternalRawInfo } from "./internal";
 import type * as Common from "@truffle/codec/common";
 import type * as Abi from "@truffle/abi-utils";
 
 export * from "./elementary"; //can't do 'export type *'
+export * from "./internal";
 
 /*
- * SECTION 1: Generic types for values in neneral (including errors).
+ * SECTION 1: Generic types for values in general (including errors).
  */
 
 /**
@@ -603,42 +605,6 @@ export interface FunctionInternalValueInfoUnknown {
   kind: "unknown";
   context: Types.ContractType;
   rawInformation: FunctionInternalRawInfo;
-}
-
-/**
- * The raw information for an internal function pointer can take two forms.
- * In non-IR Solidity, it's given as a pair of program counter values (one
- * for the constructor context and one for the deployed context).  In IR-based
- * Solidity, it's given as an index into a jump table.
- *
- * @Category Function types
- */
-export type FunctionInternalRawInfo =
-  | FunctionInternalRawInfoPcPair
-  | FunctionInternalRawInfoIndex;
-
-/**
- * In non-IR Solidity, the raw information for an internal function pointer is
- * given as pair of program counter values; one for the constructor context and
- * one for the deployed context).
- *
- * @Category Function types
- */
-export interface FunctionInternalRawInfoPcPair {
-  kind: "pcpair";
-  deployedProgramCounter: number;
-  constructorProgramCounter: number;
-}
-
-/**
- * In IR-based Solidity, the raw information for an internal function pointer is
- * given as an index into a jump table.
- *
- * @Category Function types
- */
-export interface FunctionInternalRawInfoIndex {
-  kind: "index";
-  functionIndex: number;
 }
 
 /*
